knowledge-refinery/backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.32.*
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
alembic==1.14.*
pydantic==2.10.*
pydantic-settings==2.7.*
httpx==0.28.*
python-multipart==0.0.18
python-slugify==8.0.*
pyyaml==6.0.*
beautifulsoup4==4.12.*
Pillow==11.*
Jinja2==3.1.*
